pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "floodscout"
version = "0.1.0"
description = "Historical urban flood information collection and processing framework for Weibo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7,<3",
  "python-dateutil>=2.9,<3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
  "pytest-cov>=5.0,<6",
]

[project.scripts]
floodscout = "floodscout.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
